+++ Examples/CocoaPodsTest/CocoaPodsTestTests/CocoaPodsTestTests.swift
@@ -1,0 +1,11 @@
+import Testing
+
+@testable import CocoaPodsTest
+
+struct CocoaPodsTestTests {
+
+    @Test func example() async throws {
+        // Write your test here and use APIs like `#expect(...)` to check expected conditions.
+    }
+
+}