+++ Sources/Honeycomb/Honeycomb.swift
@@ -1,11 +1,8 @@
 import BaggagePropagationProcessor
 import Foundation
-import GRPC
-import NIO
 import NetworkStatus
 import OpenTelemetryApi
 import OpenTelemetryProtocolExporterCommon
-import OpenTelemetryProtocolExporterGrpc
 import OpenTelemetryProtocolExporterHttp
 import OpenTelemetrySdk
 import ResourceExtension
@@ -14,6 +11,12 @@ import SwiftUI
 
 #if canImport(MetricKit)
     import MetricKit
+#endif
+
+#if canImport(OpenTelemetryProtocolExporterGrpc)
+    import GRPC
+    import NIO
+    import OpenTelemetryProtocolExporterGrpc
 #endif
 
 private func createAttributeDict(_ dict: [String: String]) -> [String: AttributeValue] {
@@ -75,17 +78,21 @@ public class Honeycomb {
 
         var traceExporter: SpanExporter
         if options.tracesProtocol == .grpc {
-            // Break down the URL into host and port, or use defaults from the spec.
-            let host = tracesEndpoint.host ?? "api.honeycomb.io"
-            let port = tracesEndpoint.port ?? 4317
+            #if canImport(OpenTelemetryProtocolExporterGrpc)
+                // Break down the URL into host and port, or use defaults from the spec.
+                let host = tracesEndpoint.host ?? "api.honeycomb.io"
+                let port = tracesEndpoint.port ?? 4317
 
-            let channel =
-                ClientConnection.usingPlatformAppropriateTLS(
-                    for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
-                )
-                .connect(host: host, port: port)
+                let channel =
+                    ClientConnection.usingPlatformAppropriateTLS(
+                        for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
+                    )
+                    .connect(host: host, port: port)
 
-            traceExporter = OtlpTraceExporter(channel: channel, config: otlpTracesConfig)
+                traceExporter = OtlpTraceExporter(channel: channel, config: otlpTracesConfig)
+            #else
+                throw HoneycombOptionsError.unsupportedProtocol("gRPC")
+            #endif
         } else if options.tracesProtocol == .httpJSON {
             throw HoneycombOptionsError.unsupportedProtocol("http/json")
         } else {
@@ -153,17 +160,21 @@ public class Honeycomb {
 
         var metricExporter: MetricExporter
         if options.metricsProtocol == .grpc {
-            // Break down the URL into host and port, or use defaults from the spec.
-            let host = metricsEndpoint.host ?? "api.honeycomb.io"
-            let port = metricsEndpoint.port ?? 4317
+            #if canImport(OpenTelemetryProtocolExporterGrpc)
+                // Break down the URL into host and port, or use defaults from the spec.
+                let host = metricsEndpoint.host ?? "api.honeycomb.io"
+                let port = metricsEndpoint.port ?? 4317
 
-            let channel =
-                ClientConnection.usingPlatformAppropriateTLS(
-                    for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
-                )
-                .connect(host: host, port: port)
+                let channel =
+                    ClientConnection.usingPlatformAppropriateTLS(
+                        for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
+                    )
+                    .connect(host: host, port: port)
 
-            metricExporter = OtlpMetricExporter(channel: channel, config: otlpMetricsConfig)
+                metricExporter = OtlpMetricExporter(channel: channel, config: otlpMetricsConfig)
+            #else
+                throw HoneycombOptionsError.unsupportedProtocol("gRPC")
+            #endif
         } else if options.metricsProtocol == .httpJSON {
             throw HoneycombOptionsError.unsupportedProtocol("http/json")
         } else {
@@ -187,17 +198,21 @@ public class Honeycomb {
 
         var logExporter: LogRecordExporter
         if options.logsProtocol == .grpc {
-            // Break down the URL into host and port, or use defaults from the spec.
-            let host = logsEndpoint.host ?? "api.honeycomb.io"
-            let port = logsEndpoint.port ?? 4317
+            #if canImport(OpenTelemetryProtocolExporterGrpc)
+                // Break down the URL into host and port, or use defaults from the spec.
+                let host = logsEndpoint.host ?? "api.honeycomb.io"
+                let port = logsEndpoint.port ?? 4317
 
-            let channel =
-                ClientConnection.usingPlatformAppropriateTLS(
-                    for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
-                )
-                .connect(host: host, port: port)
+                let channel =
+                    ClientConnection.usingPlatformAppropriateTLS(
+                        for: MultiThreadedEventLoopGroup(numberOfThreads: 1)
+                    )
+                    .connect(host: host, port: port)
 
-            logExporter = OtlpLogExporter(channel: channel, config: otlpLogsConfig)
+                logExporter = OtlpLogExporter(channel: channel, config: otlpLogsConfig)
+            #else
+                throw HoneycombOptionsError.unsupportedProtocol("gRPC")
+            #endif
         } else if options.logsProtocol == .httpJSON {
             throw HoneycombOptionsError.unsupportedProtocol("http/json")
         } else {