+++ CHANGELOG.md
@@ -2,10 +2,11 @@ Honeycomb OpenTelemetry SDK Changelog
 
 ## v.Next
 
-* Make `HoneycombOptions` Builder class ObjectiveC friendly
+## v 0.0.15
 
 ### New Features
 
+* Make `HoneycombOptions` Builder class ObjectiveC friendly
 * Add Cocoapods support.
 
 ### 0.0.14